requirements.txt
streamlit==1.27.0
pandas==2.2.3
python-dotenv ==1.0.0
pandas ==2.2.3
paho-mqtt==1.6.1